+++ plugins/newarc.ex/Common/Detect/tardetect.cpp
@@ -80,7 +80,7 @@ int IsTarHeader (const unsigned char *Data, unsigned int DataSize)
   size_t I;
   struct posix_header *Header;
 
-  if ((size_t)DataSize<sizeof(struct posix_header))
+  if ( DataSize<sizeof(struct posix_header))
     return -1;
 
   Header=(struct posix_header *)Data;
